fix(JavaScript): Move build files, fix template issue
