+++ clients/algoliasearch-client-javascript/packages/client-common/package.json
@@ -13,8 +13,7 @@
     "clean": "rm -rf dist/"
   },
   "engines": {
-    "node": "^14.0.0",
-    "yarn": "^3.0.0"
+    "node": "^14.0.0"
   },
   "devDependencies": {
     "@types/node": "16.11.11",
