+++ clients/algoliasearch-client-javascript/packages/client-query-suggestions/builds/browser.ts
@@ -1,10 +1,10 @@
 import type { Host, Requester } from '@algolia/client-common';
-import { HttpRequester } from '@algolia/requester-node-http';
+import { XhrRequester } from '@algolia/requester-browser-xhr';
 
-import { createQuerySuggestionsApi } from './src/querySuggestionsApi';
-import type { QuerySuggestionsApi, Region } from './src/querySuggestionsApi';
+import { createQuerySuggestionsApi } from '../src/querySuggestionsApi';
+import type { QuerySuggestionsApi, Region } from '../src/querySuggestionsApi';
 
-export * from './src/querySuggestionsApi';
+export * from '../src/querySuggestionsApi';
 
 export function querySuggestionsApi(
   appId: string,
@@ -29,12 +29,13 @@ export function querySuggestionsApi(
     apiKey,
     region,
     timeouts: {
-      connect: 2,
-      read: 5,
+      connect: 1,
+      read: 2,
       write: 30,
     },
-    requester: options?.requester ?? new HttpRequester(),
-    userAgents: [{ segment: 'Node.js', version: process.versions.node }],
+    requester: options?.requester ?? new XhrRequester(),
+    userAgents: [{ segment: 'Browser' }],
+    authMode: 'WithinQueryParameters',
     ...options,
   });
 }
