+++ openapitools.json
@@ -198,7 +198,7 @@
         "config": "#{cwd}/openapitools.json",
         "apiPackage": "src",
         "output": "#{cwd}/clients/algoliasearch-client-javascript/packages/client-predict",
-        "glob": "specs/dist/predict.yml",
+        "glob": "specs/bundled/predict.yml",
         "gitHost": "algolia",
         "gitUserId": "algolia",
         "gitRepoId": "algoliasearch-client-javascript",
