+++ scripts/post-gen/javascript.sh
@@ -7,6 +7,9 @@ FOLDER=$1
 
 # Generator does not allow new files, so we use existing ones to generate
 # our `node` and `browser` build files.
+destination=$FOLDER/builds
 
-mv $FOLDER/api.ts $FOLDER/node.ts
-mv $FOLDER/src/apis.ts $FOLDER/browser.ts
+mkdir -p $destination
+
+mv $FOLDER/api.ts $destination/node.ts
+mv $FOLDER/src/apis.ts $destination/browser.ts
