+++ templates/javascript/api-all.mustache
@@ -1,16 +1,16 @@
-{{! This file will be renamed to `browser.ts` after generating the client }}
+{{! This file will be renamed and moved to `builds/browser.ts` after generating the client }}
 
 import type { Host, Requester } from '@algolia/client-common';
 import { XhrRequester } from '@algolia/requester-browser-xhr';
 
-import { create{{capitalizedApiName}}Api } from './src/{{apiName}}Api';
-import type { {{capitalizedApiName}}Api } from './src/{{apiName}}Api';
+import { create{{capitalizedApiName}}Api } from '../src/{{apiName}}Api';
+import type { {{capitalizedApiName}}Api } from '../src/{{apiName}}Api';
 
 {{#hasRegionalHost}}
-import type { Region } from './src/{{apiName}}Api';
+import type { Region } from '../src/{{apiName}}Api';
 {{/hasRegionalHost}}
 
-export * from './src/{{apiName}}Api';
+export * from '../src/{{apiName}}Api';
 
 // eslint-disable-next-line @typescript-eslint/explicit-function-return-type
 export function {{apiName}}Api(
@@ -36,8 +36,7 @@ export function {{apiName}}Api(
 
   return create{{capitalizedApiName}}Api({
     appId,
-    apiKey,
-    {{#hasRegionalHost}}region,{{/hasRegionalHost}}
+    apiKey,{{#hasRegionalHost}}region,{{/hasRegionalHost}}
     timeouts: {
       connect: 1,
       read: 2,
