+++ templates/javascript/api.mustache
@@ -1,16 +1,16 @@
-{{! This file will be renamed to `node.ts` after generating the client }}
+{{! This file will be renamed and moved to `builds/node.ts` after generating the client }}
 
 import type { Host, Requester } from '@algolia/client-common';
 import { HttpRequester } from '@algolia/requester-node-http';
 
-import { create{{capitalizedApiName}}Api } from './src/{{apiName}}Api';
-import type { {{capitalizedApiName}}Api } from './src/{{apiName}}Api';
+import { create{{capitalizedApiName}}Api } from '../src/{{apiName}}Api';
+import type { {{capitalizedApiName}}Api } from '../src/{{apiName}}Api';
 
 {{#hasRegionalHost}}
-import type { Region } from './src/{{apiName}}Api';
+import type { Region } from '../src/{{apiName}}Api';
 {{/hasRegionalHost}}
 
-export * from './src/{{apiName}}Api';
+export * from '../src/{{apiName}}Api';
 
 export function {{apiName}}Api(
   appId: string,
@@ -35,8 +35,7 @@ export function {{apiName}}Api(
 
   return create{{capitalizedApiName}}Api({
     appId,
-    apiKey,
-    {{#hasRegionalHost}}region,{{/hasRegionalHost}}
+    apiKey,{{#hasRegionalHost}}region,{{/hasRegionalHost}}
     timeouts: {
       connect: 2,
       read: 5,
