+++ templates/javascript/tsconfig.mustache
@@ -3,6 +3,6 @@
   "compilerOptions": {
     "outDir": "dist",
   },
-  "include": ["{{apiPackage}}", "model", "node.ts", "browser.ts"],
+  "include": ["{{apiPackage}}", "model", "builds/node.ts", "builds/browser.ts"],
   "exclude": ["dist", "node_modules"]
 }
